+++ apps/sv/src/main/scala/com/daml/network/sv/config/package.scala
@@ -1,7 +1,0 @@
-package com.daml.network.sv
-
-import com.daml.network.config.GlobalOnlyDomainConfig
-
-package object config {
-  type SvDomainConfig = GlobalOnlyDomainConfig
-}