+++ cluster/pulumi/canton-network/src/validator.ts
@@ -62,6 +62,7 @@ export async function installValidator(
         'cn-node-0.1.0-SNAPSHOT/dars/directory-service-0.1.0.dar',
         'cn-node-0.1.0-SNAPSHOT/dars/splitwell-0.1.0.dar',
       ],
+      globalDomainUrl: 'http://global-domain.svc:5008',
     },
     dependsOn
   );